# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CIMOL API with Golang.\nGolang memiliki keunggulan dalam skalabilitas horizontal dan pemeliharaan kode yang lebih sederhana,\nGolang juga memiliki keamanan bawaan dan memudahkan pencegahan beberapa jenis serangan, serta\nkinerja yang signifikan dalam kasus pengolahan beban kerja tinggi dan pembuatan API yang membutuhkan respons cepat.",
        "title": "CIMOL API with Golang",
        "contact": {
            "name": "API CNAF MOBILE",
            "url": "https://www.cnaf.co.id/",
            "email": "dimas.zikri@cnaf.co.id"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api",
    "paths": {
...
        "/tools/FindDataAssignFC": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DataAssignFC"
                ],
                "summary": "Find Data Assign FC",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FDataAssignFC"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.FindDataAssignFC"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/FindDataAssignFcHistory": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DataAssignFC"
                ],
                "summary": "Find Data Assign FC History",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FDataAssignFC"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.FindDataAssignFcHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/GetCollectionHistory": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CollectionHistory"
                ],
                "summary": "Get Collection History",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetCollection"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetCollectionHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/GetCustomerCard": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CustomerCard"
                ],
                "summary": "GetCustomerCard",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CustomerCard"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetCustomerCard"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/GetDataAssignFC": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DataAssignFC"
                ],
                "summary": "Get Data Assign FC",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataAssignFC"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetDataAssignFC"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/GetDataAssignFcHistory": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DataAssignFC"
                ],
                "summary": "Get Data Assign FC History",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataAssignFC"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetDataAssignFcHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/GetLovGroup": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "LovGroup"
                ],
                "summary": "Get LovGroup",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataAssignFC"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetLovGroup"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/tools/GetTblDataColl": {
            "post": {
                "description": "OAuth protects our entity endpoints",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DataColl"
                ],
                "summary": "Get Data Coll",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataColl"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token get from login",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetTblDataColl"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helpers.ResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "responsecode": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helpers.ResponseWithoutData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "responsecode": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.FDataAssignFC": {
            "type": "object",
            "properties": {
                "gps": {
                    "type": "string",
                    "example": "0;0"
                },
                "orderby": {
                    "type": "string",
                    "example": "applicationid"
                },
                "parameter": {
                    "type": "string",
                    "example": "-1"
                },
                "searchdata": {
                    "type": "string",
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "example": "1879"
                }
            }
        },
...
        "models.GetCollection": {
            "type": "object",
            "properties": {
                "applicationid": {
                    "type": "string"
                },
                "customerid": {
                    "type": "string"
                }
            }
        },
...
    }
}